+++ ricaun.NUnit.Tests/NUnitTests.cs
@@ -172,7 +172,8 @@ namespace ricaun.NUnit.Tests
         [TestCase("*.TestName?", 2)]
         [TestCase("*.TestCases(?)", 2)]
         [TestCase("*.TestSame?", 2)]
-        [TestCase("*", 32)]
+        [TestCase("*.TestTask*", 5)]
+        [TestCase("*", 37)]
         public void TestAssembly_Filter(string testName, int numberOfTests)
         {
             TestEngineFilter.Add(testName);