+++ ricaun.NUnit/Services/ActivatorService.cs
@@ -56,62 +56,48 @@ namespace ricaun.NUnit.Services
 
             var methodParams = GetMethodOrderParameters(method, possibleParams);
 
-            if (method.ReturnType == typeof(Task))
+            if (IsReturnTypeEqualsTask(method))
             {
+                var taskInvoke = method.Invoke(obj, methodParams);
+
+                if (taskInvoke is Task task)
+                {
+                    return InvokeTask(task);
+                }
+
                 throw new TaskCanceledException("Task method not supported!");
-
-                //var taskInvoke = (Task)method.Invoke(obj, methodParams);
-                //if (!taskInvoke.Wait(2000))
-                //{
-                //    Console.WriteLine("Task 2000ms Timeout");
-                //};
-
-                //var task = Task.Run(async () =>
-                //{
-                //    await taskInvoke;
-                //});
-                //task.GetAwaiter().GetResult();
-
-                //return;
             }
 
             return method.Invoke(obj, methodParams);
         }
 
         /// <summary>
-        /// InvokeAsync
+        /// Invoke Task and Return 'Result'
         /// </summary>
-        /// <param name="obj"></param>
-        /// <param name="method"></param>
-        /// <param name="possibleParams"></param>
+        /// <param name="task"></param>
         /// <returns></returns>
-        public async Task InvokeAsync(object obj, MethodInfo method, params object[] possibleParams)
+        private object InvokeTask(Task task)
         {
-            if (method is null)
-                return;
-
-            var methodParams = GetMethodOrderParameters(method, possibleParams);
-
-            if (method.ReturnType == typeof(Task))
-            {
-                await (Task)method.Invoke(obj, methodParams);
-                return;
-            }
-
-            method.Invoke(obj, methodParams);
+            Task.Run(async () => { await task; }).GetAwaiter().GetResult();
+            var resultProperty = task.GetType().GetProperty("Result");
+            return resultProperty?.GetValue(task);
         }
 
         /// <summary>
-        /// InvokeAsync
+        /// Is <paramref name="method"/> Task or Task<>
         /// </summary>
-        /// <param name="obj"></param>
-        /// <param name="methodName"></param>
-        /// <param name="possibleParams"></param>
+        /// <param name="method"></param>
         /// <returns></returns>
-        public async Task InvokeAsync(object obj, string methodName, params object[] possibleParams)
+        private bool IsReturnTypeEqualsTask(MethodInfo method)
         {
-            var method = obj?.GetType().GetMethod(methodName);
-            await InvokeAsync(obj, method, possibleParams);
+            var returnType = method.ReturnType;
+
+            if (returnType.IsGenericType && returnType.GetGenericTypeDefinition() == typeof(Task<>))
+            {
+                return true;
+            }
+
+            return returnType == typeof(Task);
         }
         #endregion
 
@@ -129,13 +115,22 @@ namespace ricaun.NUnit.Services
 
             foreach (ParameterInfo parameter in methodBase.GetParameters())
             {
-                object o = possibleParamsTemp.FirstOrDefault(e => e.GetType().Equals(parameter.ParameterType));
+                object o = possibleParamsTemp.FirstOrDefault(e => IsParameterTypeSimilar(parameter, e));
                 possibleParamsTemp.Remove(o);
                 result.Add(o);
             }
 
             return result.ToArray();
         }
+
+        private bool IsParameterTypeSimilar(ParameterInfo parameter, object parameterValue)
+        {
+            if (parameter.ParameterType == typeof(object))
+                return true;
+
+            return parameterValue.GetType().Equals(parameter.ParameterType);
+        }
+
         #endregion
     }
 }