+++ CHANGELOG.md
@@ -5,8 +5,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.2.2] / 2023-03-14
+### Features
+- [x] Enable Task Tests to work with the Framework
+- [] Async Test Timeout
 ### Updated
+- Update `TestModel`
 - Update `AppDomainUtils`
+- Add `TaskTests`
+- Add `InvokeTask`
+- Add `IsReturnTypeEqualsTask`
+- Add `IsParameterTypeSimilar`
 
 ## [1.2.1] / 2023-02-15
 ### Features